refactoring and comment
