+++ Wire-iOS Tests/GiphySearchViewControllerTests.swift
@@ -27,10 +27,7 @@ final class GiphySearchViewControllerTests: XCTestCase {
     override func setUp() {
         super.setUp()
 
-        mockConversation = MockConversation()
-        mockConversation.conversationType = .oneOnOne
-        mockConversation.displayName = "John Doe"
-        mockConversation.connectedUser = MockUser.mockUsers().last!
+        mockConversation = MockConversation.onoOnOneConversation()
     }
     
     override func tearDown() {
@@ -57,54 +54,4 @@ final class GiphySearchViewControllerTests: XCTestCase {
         XCTAssertNil(sut)
     }
 
-}
-
-final class GiphySnapshotTests: ZMSnapshotTestCase {
-    var sut: GiphySearchViewController!
-
-    var mockConversation: MockConversation!
-    var mockNavigationController: UINavigationController!
-
-    override func setUp() {
-        super.setUp()
-
-        mockConversation = MockConversation()
-        mockConversation.conversationType = .oneOnOne ///TODO: mv to class method
-        mockConversation.displayName = "John Doe"
-        mockConversation.connectedUser = MockUser.mockUsers().last!
-
-        let searchTerm: String = "apple"
-        sut = GiphySearchViewController(withSearchTerm: searchTerm, conversation: (mockConversation as Any) as! ZMConversation)
-        mockNavigationController = sut.wrapInsideNavigationController()
-
-        sut.collectionView?.backgroundColor = .white
-
-        UIView.setAnimationsEnabled(false)
-
-        recordMode = true
-    }
-
-    override func tearDown() {
-        sut = nil
-        mockConversation = nil
-        mockNavigationController = nil
-
-        UIView.setAnimationsEnabled(true)
-
-        super.tearDown()
-    }
-
-    func testEmptySearchScreenWithKeyword(){
-        verify(view: mockNavigationController.view)
-    }
-
-    func testConfirmationScreenWithDisabledSendButton(){
-        let data = self.data(forResource: "animated", extension: "gif")!
-        let image = FLAnimatedImage(animatedGIFData: data)
-
-        let confirmationController = sut.pushConfirmationViewController(ziph: nil, previewImage: image, animated: false)
-        confirmationController.view.backgroundColor = .white
-
-        verify(view: mockNavigationController.view)
-    }
 }
